# backend/contracts/rustrag.openapi.yaml
openapi: 3.1.0
info:
  title: RustRAG API
  version: 0.1.0
  description: |
    Runtime API contract for the current RustRAG backend implementation.

    This contract is hand-maintained from the implemented Axum routes in `backend/src/interfaces/http`.
    It aims to reflect the currently available endpoints, request bodies, response shapes,
    auth requirements, and common error envelopes as accurately as practical.
servers:
  - url: http://localhost:8095
    description: Local default
security:
  - bearerAuth: []
...
paths:
...
  /v1/retrieval-runs:
    get:
      tags: [retrieval]
      summary: List retrieval runs
      operationId: listRetrievalRuns
      security: []
      parameters:
        - $ref: '#/components/parameters/projectIdOptional'
      responses:
        '200':
          description: Retrieval run list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrievalRunSummary'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [retrieval]
      summary: Create a retrieval run record manually
      operationId: createRetrievalRun
      description: Auth header is currently required by the handler, but no scope check is enforced.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRetrievalRunRequest'
      responses:
        '200':
          description: Created retrieval run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalRunSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/query:
    post:
      tags: [retrieval]
      summary: Run live retrieval + generation query
      operationId: runQuery
      description: |
        Requires `query:run`, `workspace:admin`, or an `instance_admin` token.

        The backend combines lexical matching with optional semantic matching based on stored chunk embeddings,
        then calls the configured LLM gateway to produce the final answer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token minted by `POST /v1/auth/tokens`.
        Note: some read/list endpoints are currently public and do not require auth.
  parameters:
...
    projectIdOptional:
      in: query
      name: project_id
      required: false
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
          examples:
            apiErrorEnvelope:
              value:
                error: unauthorized
        text/plain:
          schema:
            type: string
          examples:
            authExtractorError:
              value: missing authorization header
      headers:
        WWW-Authenticate:
          required: false
          schema:
            type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
        text/plain:
          schema:
            type: string
  schemas:
    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RetrievalRunSummary:
      type: object
      required: [id, project_id, query_text, top_k]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        query_text:
          type: string
        model_profile_id:
          type: [string, 'null']
          format: uuid
        top_k:
          type: integer
          format: int32
        response_text:
          type: [string, 'null']
    CreateRetrievalRunRequest:
      type: object
      required: [project_id, query_text]
      properties:
        project_id:
          type: string
          format: uuid
        query_text:
          type: string
        model_profile_id:
          type: [string, 'null']
          format: uuid
        top_k:
          type: [integer, 'null']
          format: int32
        response_text:
          type: [string, 'null']
    QueryRequest:
      type: object
      required: [project_id, query_text]
      properties:
        project_id:
          type: string
          format: uuid
        query_text:
          type: string
        model_profile_id:
          type: [string, 'null']
          format: uuid
        embedding_model_profile_id:
          type: [string, 'null']
          format: uuid
        top_k:
          type: [integer, 'null']
          format: int32
    QueryResponse:
      type: object
      required: [retrieval_run_id, project_id, answer, references, mode]
      properties:
        retrieval_run_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        answer:
          type: string
        references:
          type: array
          items:
            type: string
          example:
            - document:01958f52-1b41-7e1a-81f0-68bb2ac6df54:chunk:0
        mode:
          type: string
          example: gateway_live
...